Limit number of registers collected
In order to compare the fast tracepoints with LTTng, I reduced to
8 the number of registers that is collected with "collect $regs".

This branch uses the standard fast tracepoint to collect data. The
number of register that are saved in collect $regs was changed to
ensure proper comparaison with the LTTng branch.

This is done because LTTng save every tracepoints instead of
erasing them like GDB does after its buffer is full. Because the
tests were done using high-frequency events, LTTng traces could
reach up to 4 GB. Thus, only 64 bytes is now collected for testing
purposes.